refactor dpm command center exception routes
