+++ src/app/routers/dpm_command_center_monitoring.py
@@ -3,27 +3,17 @@ from fastapi import APIRouter, Path, Query
 from app.contracts.dpm_command_center import (
     DpmCommandCenterForwardRequest,
     DpmCommandCenterGatewayResponse,
-    DpmCommandCenterResolveExceptionRequest,
-    DpmExceptionSummaryGatewayResponse,
-    DpmExceptionSummaryRequest,
-    DpmOutcomeReviewErrorDetail,
 )
 from app.middleware.correlation import correlation_id_var
-from app.routers.dpm_openapi import manage_upstream_error_responses
-from app.services.dpm_service_provider import dpm_command_center_service
-
-_UPSTREAM_ERROR_RESPONSES = manage_upstream_error_responses(
-    error_model=DpmOutcomeReviewErrorDetail,
-    not_found_description="lotus-manage could not find the requested command-center resource.",
-    conflict_description="lotus-manage rejected the command-center request as conflicting.",
-    invalid_payload_description="lotus-manage rejected the command-center payload as invalid.",
-    unavailable_description="lotus-manage command-center authority is unavailable or degraded.",
+from app.routers.dpm_command_center_monitoring_common import (
+    UPSTREAM_COMMAND_CENTER_ERROR_RESPONSES,
 )
+from app.services.dpm_service_provider import dpm_command_center_service
 
 router = APIRouter(
     prefix="/api/v1/dpm/command-center",
     tags=["DPM Command Center"],
-    responses=_UPSTREAM_ERROR_RESPONSES,
+    responses=UPSTREAM_COMMAND_CENTER_ERROR_RESPONSES,
 )
 
 
@@ -91,92 +81,5 @@ async def get_monitoring_run(
 ) -> DpmCommandCenterGatewayResponse:
     return await dpm_command_center_service().get_monitoring_run(
         monitoring_run_id=monitoring_run_id,
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/exceptions",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="List DPM monitoring exceptions",
-    description=(
-        "What: lists manage-owned mandate monitoring exceptions. When: use this for Workbench "
-        "attention queues and operations triage by mandate, portfolio, or state. How: Gateway "
-        "preserves manage exception ids, severity, reason codes, state, and recommended action."
-    ),
-)
-async def list_monitoring_exceptions(
-    mandate_id: str | None = Query(default=None, description="Optional mandate id filter."),
-    portfolio_id: str | None = Query(default=None, description="Optional portfolio id filter."),
-    state: str | None = Query(
-        default=None,
-        description="Optional manage-published exception state filter.",
-        examples=["ACTIVE"],
-    ),
-    limit: int = Query(default=50, ge=1, le=200, description="Maximum exceptions to return."),
-    cursor: str | None = Query(default=None, description="Cursor from a previous page."),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().list_monitoring_exceptions(
-        filters={
-            "mandate_id": mandate_id,
-            "portfolio_id": portfolio_id,
-            "state": state,
-            "limit": limit,
-            "cursor": cursor,
-        },
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/exceptions/{exception_id}/resolve",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="Resolve DPM monitoring exception",
-    description=(
-        "What: forwards an exception resolution reason to lotus-manage. When: use this after a "
-        "PM, supervisor, or operator has reviewed the exception. How: Gateway does not close "
-        "exceptions locally; it returns the manage-owned resolved exception payload."
-    ),
-)
-async def resolve_monitoring_exception(
-    request: DpmCommandCenterResolveExceptionRequest,
-    exception_id: str = Path(
-        ...,
-        description="Manage-owned monitoring exception identifier.",
-        examples=["me_source_readiness_001"],
-    ),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().resolve_monitoring_exception(
-        exception_id=exception_id,
-        body={"resolution_reason": request.resolution_reason},
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.post(
-    "/exceptions/{exception_id}/ai-summary",
-    response_model=DpmExceptionSummaryGatewayResponse,
-    summary="Request DPM exception AI summary",
-    description=(
-        "What: requests a governed lotus-ai exception-summary workflow-pack run from "
-        "manage-owned monitoring exception evidence. When: call this only for internal PM, "
-        "investment-control, or operations triage after the exception is visible in the command "
-        "center. How: Gateway reads the manage exception queue, builds a bounded no-raw-payload "
-        "evidence envelope for the selected exception, then executes lotus-ai "
-        "dpm_exception_summary.pack@v1 as lotus-gateway; Gateway does not generate narrative, "
-        "score PMs, approve trades, contact clients, route orders, or invent evidence."
-    ),
-)
-async def request_exception_summary(
-    request: DpmExceptionSummaryRequest,
-    exception_id: str = Path(
-        ...,
-        description="Manage-owned monitoring exception identifier for the bounded AI handoff.",
-        examples=["me_source_readiness_001"],
-    ),
-) -> DpmExceptionSummaryGatewayResponse:
-    return await dpm_command_center_service().request_exception_summary(
-        exception_id=exception_id,
-        request=request,
         correlation_id=correlation_id_var.get(),
     )
